+++ src/controllers/DefaultController.php
@@ -14,10 +14,10 @@ use Craft;
 use craft\errors\AssetDisallowedExtensionException;
 use craft\helpers\Path as PathHelper;
 use craft\web\Controller;
-use craft\web\Response;
 use nystudio107\transcoder\Transcoder;
 use yii\base\ExitException;
 use yii\web\BadRequestHttpException;
+use yii\web\Response;
 
 use function count;
 use function is_array;
@@ -33,7 +33,7 @@ class DefaultController extends Controller
     // =========================================================================
 
     /**
-     * @var    bool|array Allows anonymous access to this controller's actions.
+     * @var    bool|int|array<int|string> Allows anonymous access to this controller's actions.
      *         The actions must be in 'kebab-case'
      * @access protected
      */