+++ tests/craft5/config/app.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'components' => [
+        'projectConfig' => [
+            'class' => craft\services\ProjectConfig::class,
+            'writeYamlAutomatically' => false,
+        ],
+    ],
+];