+++ tests/craft5/config/general.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'runQueueAutomatically' => false,
+    'aliases' => [
+        '@webroot' => sys_get_temp_dir() . '/transcoder-craft5-test/web',
+        '@web' => 'https://transcoder.test',
+    ],
+];